--- packages/electron-rpc/src/main/index.ts
@@ -1,3 +1,0 @@
-export * from './main-channel';
-export * from './main-invoker';
-export * from './main-receiver';
