--- packages/electron-rpc/src/renderer/index.ts
@@ -1,3 +1,0 @@
-export * from './renderer-channel';
-export * from './renderer-invoker';
-export * from './renderer-receiver';
